Moved content div
div with class content was wrongly placed in the navbar div. This resulted in the "This website is awesome text" to be placed next to the header links because of the flex direction being row. I moved the div with the class of content outside of navbar but still inside of first-section div and it was fixed.